Add some helper functions for getting settings
Add 4 helper functions to get settings in different types.
Change the way json is imported in utils, use if, instead of a try block
Change retries variable in utilities to be read every traktJsonRequest,
instead of on plugin load, this way a change in that setting doesn't
need a plugin restart.