Add some error checking in loading XBMC shows.
Check for empty data, and abort tv show sync if it encounters, since this means there was a problem with the json result.
Empty data is not the same as the episodes key being missing, as there will still be data in the result.
Check for 'episodes' in JSON data, if it doesn't exist, continue on loading the next show.