Add addtolist and removefrom list actions
Add addtolist action, pass along list=, and optional media_type=<movie/show>, dbid=<#> to manually add an item to a list, if the list does not exist, it will be created.
Add removefrom action, pass along list=, and optional media_type=<movie/show>, dbid=<#> to manually remove an item from a list.
Fix argument parsing to no longer force values to lower case.
Add a wrapper function to check if a list is on trakt, and utilize a 10 minute caching
Add a wrapper function to get the slug of a trakt list