Add ability to edit list within managelists dialog
Edit managelists dialog to include controls for editing a list, these will be visible once a list is selected.
Added ability to hide a list in the itemslist dialog, useful for if you have some lists that you just don't want to use in XBMC.
managelists dialog remembers your list selection when returning to the from editing a list.