Fix: Sync progress to library when no runtime info
TVDB does not provide runtime metadata for episodes. If this can not then be
filled by Kodi introspecting the file for the item (e.g. library item that is
a .STRM file) the info label will be `0`. Meaning the resume position set by
Trackt will be `0` as the calculation is `runtime * progress_pct`.

Episode runtime metadata exists in the metadata provided by Trakt for episodes
so this can be queried in these cases to calculate a resume time that is
reasonably accurate.

This depends on fuzeman/trakt.py#69 being merged and propagated to
Razzeee/script.module.trakt