Fix id handling for manual actions
We ignored the correct ids in some cases and ended up thinking
it was a tvdb id, even if we knew better.

Rename parseIdToTraktIds to guessBestTraktId as it's clearer on what it
actially does.  It should only be used, if we don't know which kind
of id we're handling, otherwise it will lead us to loose id types.

Fixes https://github.com/trakt/script.trakt/issues/530