settings: Migrate to new settings format
The current way of settings for trakt is deprecated and should migrate
to the new format (see [0]). This commit performs the migration to the
new settings format.

Everything is kept as close as possible on how it was, with all standard
configuration settings added to the 'standard' level, auth settings to
the 'basic' level (as otherwise the addon is not that useful) and some
more complex things (debug for example) to higher levels.

The 'internal settings' that are used to store some tokens etc, are at
the 'never show in GUI' level 4.

[0]: https://kodi.wiki/view/Add-on_settings

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>